consolidate some code
