+++ packages/mds-metrics-sheet/vehicle-counts.ts
@@ -13,58 +13,13 @@
     See the License for the specific language governing permissions and
     limitations under the License.
  */
-
-import GoogleSpreadsheet from 'google-spreadsheet'
-
-import { promisify } from 'util'
-
 import requestPromise from 'request-promise'
 
 import log from '@mds-core/mds-logger'
-import {
-  JUMP_PROVIDER_ID,
-  LIME_PROVIDER_ID,
-  BIRD_PROVIDER_ID,
-  LYFT_PROVIDER_ID,
-  WHEELS_PROVIDER_ID,
-  SPIN_PROVIDER_ID,
-  SHERPA_LA_PROVIDER_ID,
-  BOLT_PROVIDER_ID
-} from '@mds-core/mds-providers'
 
 import { VehicleCountResponse, VehicleCountRow } from './types'
-
-// The list of providers ids on which to report
-const reportProviders = [
-  JUMP_PROVIDER_ID,
-  LIME_PROVIDER_ID,
-  BIRD_PROVIDER_ID,
-  LYFT_PROVIDER_ID,
-  WHEELS_PROVIDER_ID,
-  SPIN_PROVIDER_ID,
-  SHERPA_LA_PROVIDER_ID,
-  BOLT_PROVIDER_ID
-]
-
-const creds = {
-  client_email: process.env.GOOGLE_CLIENT_EMAIL,
-  private_key: process.env.GOOGLE_PRIVATE_KEY ? process.env.GOOGLE_PRIVATE_KEY.split('\\n').join('\n') : null
-}
-
-async function appendSheet(sheetName: string, rows: ({ date: string; name: string } & unknown)[]) {
-  const doc = new GoogleSpreadsheet(process.env.SPREADSHEET_ID)
-  await promisify(doc.useServiceAccountAuth)(creds)
-  const info = await promisify(doc.getInfo)()
-  log.info(`Loaded doc: ${info.title} by ${info.author.email}`)
-  const sheet = info.worksheets.filter((s: { title: string; rowCount: number } & unknown) => s.title === sheetName)[0]
-  log.info(`${sheetName} sheet: ${sheet.title} ${sheet.rowCount}x${sheet.colCount}`)
-  if (sheet.title === sheetName) {
-    const inserted = rows.map(insert_row => promisify(sheet.addRow)(insert_row))
-    log.info(`Wrote ${inserted.length} rows.`)
-    return Promise.all(inserted)
-  }
-  log.info('Wrong sheet!')
-}
+import { appendSheet } from './metrics-log-utils'
+import { reportProviders, getAuthToken } from './shared-utils'
 
 export function sumColumns(keysToSummarize: string[], row: VehicleCountRow) {
   return keysToSummarize.reduce((acc, veniceAreaKey) => acc + row.areas_48h[veniceAreaKey] || 0, 0)
@@ -93,21 +48,8 @@ async function getProviderMetrics(iter: number): Promise<({ date: string; name: 
     throw new Error(`Failed to write to sheet after 10 tries!`)
   }
 
-  const token_options = {
-    method: 'POST',
-    url: `${process.env.AUTH0_DOMAIN}/oauth/token`,
-    headers: { 'content-type': 'application/json' },
-    body: {
-      grant_type: 'client_credentials',
-      client_id: process.env.CLIENT_ID,
-      client_secret: process.env.CLIENT_SECRET,
-      audience: process.env.AUDIENCE
-    },
-    json: true
-  }
-
   try {
-    const token = await requestPromise(token_options)
+    const token = await getAuthToken()
     const counts_options = {
       uri: 'https://api.ladot.io/daily/admin/vehicle_counts',
       headers: { authorization: `Bearer ${token.access_token}` },
